Introduce `BaseApp` and message
Make message and recipient types standard. This allows to extract a
`BaseApp` generic and to re-use the same logic for all Bots implementing
the `MessagingBot` interface.
