+++ access/common/app.go
@@ -1,0 +1,439 @@
+package common
+
+import (
+	"context"
+	pd "github.com/gravitational/teleport-plugins/lib/plugindata"
+	"github.com/gravitational/teleport-plugins/lib/stringset"
+	"github.com/gravitational/teleport-plugins/lib/watcherjob"
+	"github.com/gravitational/teleport/api/types"
+	grpcbackoff "google.golang.org/grpc/backoff"
+	"time"
+
+	"github.com/gravitational/teleport-plugins/lib"
+	"github.com/gravitational/teleport-plugins/lib/logger"
+	"github.com/gravitational/teleport/api/client"
+	"github.com/gravitational/teleport/api/client/proto"
+	"github.com/gravitational/trace"
+	"google.golang.org/grpc"
+)
+
+const (
+	// minServerVersion is the minimal teleport version the plugin supports.
+	minServerVersion = "6.1.0-beta.1"
+	// grpcBackoffMaxDelay is a maximum time GRPC client waits before reconnection attempt.
+	grpcBackoffMaxDelay = time.Second * 2
+	// InitTimeout is used to bound execution time of health check and teleport version check.
+	initTimeout = time.Second * 10
+	// handlerTimeout is used to bound the execution time of watcher event handler.
+	handlerTimeout = time.Second * 5
+)
+
+type BaseApp[T PluginConfiguration] struct {
+	PluginName string
+	ApiClient  *client.Client
+	Bot        MessagingBot
+	MainJob    lib.ServiceJob
+	pd         *pd.CompareAndSwap[GenericPluginData]
+	Conf       T
+	NewBot     func(T, string, string) (MessagingBot, error)
+
+	*lib.Process
+}
+
+// NewApp creates a new base app and initialize its main job
+func NewApp[T PluginConfiguration](conf T, pluginName string, newBot BotFactory[T]) *BaseApp[T] {
+	app := BaseApp[T]{
+		PluginName: pluginName,
+		Conf:       conf,
+		NewBot:     newBot,
+	}
+	app.MainJob = lib.NewServiceJob(app.run)
+	return &app
+}
+
+// Run initializes and runs a watcher and a callback server
+func (a *BaseApp[T]) Run(ctx context.Context) error {
+	// Initialize the process.
+	a.Process = lib.NewProcess(ctx)
+	a.SpawnCriticalJob(a.MainJob)
+	<-a.Process.Done()
+	return a.Err()
+}
+
+// Err returns the error app finished with.
+func (a *BaseApp[T]) Err() error {
+	return trace.Wrap(a.MainJob.Err())
+}
+
+// WaitReady waits for http and watcher service to start up.
+func (a *BaseApp[T]) WaitReady(ctx context.Context) (bool, error) {
+	return a.MainJob.WaitReady(ctx)
+}
+
+func (a *BaseApp[T]) checkTeleportVersion(ctx context.Context) (proto.PingResponse, error) {
+	log := logger.Get(ctx)
+	log.Debug("Checking Teleport server version")
+	pong, err := a.ApiClient.WithCallOptions(grpc.WaitForReady(true)).Ping(ctx)
+	if err != nil {
+		if trace.IsNotImplemented(err) {
+			return pong, trace.Wrap(err, "server version must be at least %s", minServerVersion)
+		}
+		log.Error("Unable to get Teleport server version")
+		return pong, trace.Wrap(err)
+	}
+	err = lib.AssertServerVersion(pong, minServerVersion)
+	return pong, trace.Wrap(err)
+}
+
+func (a *BaseApp[T]) initTeleport(ctx context.Context, conf PluginConfiguration) (clusterName, webProxyAddr string, err error) {
+	var (
+		pong proto.PingResponse
+	)
+
+	bk := grpcbackoff.DefaultConfig
+	bk.MaxDelay = grpcBackoffMaxDelay
+	if a.ApiClient, err = client.New(ctx, client.Config{
+		Addrs:       conf.GetTeleportConfig().GetAddrs(),
+		Credentials: conf.GetTeleportConfig().Credentials(),
+		DialOpts: []grpc.DialOption{
+			grpc.WithConnectParams(grpc.ConnectParams{Backoff: bk, MinConnectTimeout: initTimeout}),
+			grpc.WithReturnConnectionError(),
+		},
+	}); err != nil {
+		return "", "", trace.Wrap(err)
+	}
+
+	if pong, err = a.checkTeleportVersion(ctx); err != nil {
+		return "", "", trace.Wrap(err)
+	}
+
+	if pong.ServerFeatures.AdvancedAccessWorkflows {
+		webProxyAddr = pong.ProxyPublicAddr
+	}
+
+	return pong.ClusterName, webProxyAddr, nil
+}
+
+func (a *BaseApp[T]) OnWatcherEvent(ctx context.Context, event types.Event) error {
+	if kind := event.Resource.GetKind(); kind != types.KindAccessRequest {
+		return trace.Errorf("unexpected kind %s", kind)
+	}
+	op := event.Type
+	reqID := event.Resource.GetName()
+	ctx, _ = logger.WithField(ctx, "request_id", reqID)
+
+	switch op {
+	case types.OpPut:
+		ctx, _ = logger.WithField(ctx, "request_op", "put")
+		req, ok := event.Resource.(types.AccessRequest)
+		if !ok {
+			return trace.Errorf("unexpected resource type %T", event.Resource)
+		}
+		ctx, log := logger.WithField(ctx, "request_state", req.GetState().String())
+
+		var err error
+		switch {
+		case req.GetState().IsPending():
+			err = a.onPendingRequest(ctx, req)
+		case req.GetState().IsApproved():
+			err = a.onResolvedRequest(ctx, req)
+		case req.GetState().IsDenied():
+			err = a.onResolvedRequest(ctx, req)
+		default:
+			log.WithField("event", event).Warn("Unknown request state")
+			return nil
+		}
+
+		if err != nil {
+			log.WithError(err).Errorf("Failed to process request")
+			return trace.Wrap(err)
+		}
+
+		return nil
+	case types.OpDelete:
+		ctx, log := logger.WithField(ctx, "request_op", "delete")
+
+		if err := a.onDeletedRequest(ctx, reqID); err != nil {
+			log.WithError(err).Errorf("Failed to process deleted request")
+			return trace.Wrap(err)
+		}
+		return nil
+	default:
+		return trace.BadParameter("unexpected event operation %s", op)
+	}
+}
+
+func (a *BaseApp[T]) run(ctx context.Context) error {
+	var err error
+
+	log := logger.Get(ctx)
+
+	if err = a.init(ctx); err != nil {
+		return trace.Wrap(err)
+	}
+	watcherJob := watcherjob.NewJob(
+		a.ApiClient,
+		watcherjob.Config{
+			Watch:            types.Watch{Kinds: []types.WatchKind{{Kind: types.KindAccessRequest}}},
+			EventFuncTimeout: handlerTimeout,
+		},
+		a.OnWatcherEvent,
+	)
+	a.SpawnCriticalJob(watcherJob)
+	ok, err := watcherJob.WaitReady(ctx)
+	if err != nil {
+		return trace.Wrap(err)
+	}
+
+	a.MainJob.SetReady(ok)
+	if ok {
+		log.Info("Plugin is ready")
+	} else {
+		log.Error("Plugin is not ready")
+	}
+
+	<-watcherJob.Done()
+
+	return trace.Wrap(watcherJob.Err())
+}
+
+func (a *BaseApp[T]) init(ctx context.Context) error {
+	ctx, cancel := context.WithTimeout(ctx, initTimeout)
+	defer cancel()
+	log := logger.Get(ctx)
+
+	clusterName, webProxyAddr, err := a.initTeleport(ctx, a.Conf)
+	if err != nil {
+		return trace.Wrap(err)
+	}
+
+	a.Bot, err = a.NewBot(a.Conf, clusterName, webProxyAddr)
+	if err != nil {
+		return trace.Wrap(err)
+	}
+
+	a.pd = pd.NewCAS(
+		a.ApiClient,
+		a.PluginName,
+		types.KindAccessRequest,
+		EncodePluginData,
+		DecodePluginData,
+	)
+
+	log.Debug("Starting API health check...")
+	if err = a.Bot.HealthCheck(ctx); err != nil {
+		return trace.Wrap(err, "API health check failed")
+	}
+
+	log.Debug("API health check finished ok")
+	return nil
+}
+
+func (a *BaseApp[T]) onPendingRequest(ctx context.Context, req types.AccessRequest) error {
+	log := logger.Get(ctx)
+
+	reqID := req.GetName()
+	reqData := pd.AccessRequestData{
+		User:          req.GetUser(),
+		Roles:         req.GetRoles(),
+		RequestReason: req.GetRequestReason(),
+	}
+
+	_, err := a.pd.Create(ctx, reqID, GenericPluginData{AccessRequestData: reqData})
+
+	if !trace.IsAlreadyExists(err) {
+		if err != nil {
+			return trace.Wrap(err)
+		}
+
+		// This is a new access request, we should create new messages
+		if channels := a.getMessageRecipients(ctx, req); len(channels) > 0 {
+			if err := a.broadcastMessages(ctx, channels, reqID, reqData); err != nil {
+				return trace.Wrap(err)
+			}
+		} else {
+			log.Warning("No channel to post")
+		}
+	}
+
+	// This is an already existing access request, we post reviews and update its status
+	if reqReviews := req.GetReviews(); len(reqReviews) > 0 {
+		if err := a.postReviewReplies(ctx, reqID, reqReviews); err != nil {
+			return trace.Wrap(err)
+		}
+
+		err := a.updateMessages(ctx, reqID, pd.Unresolved, "", reqReviews)
+		if err != nil {
+			return trace.Wrap(err)
+		}
+	}
+
+	return nil
+}
+
+func (a *BaseApp[T]) onResolvedRequest(ctx context.Context, req types.AccessRequest) error {
+	var replyErr error
+
+	// We always post review replies in thread. If the messaging service does not support
+	// threading this will do nothing
+	if err := a.postReviewReplies(ctx, req.GetName(), req.GetReviews()); err != nil {
+		replyErr = trace.Wrap(err)
+
+	}
+
+	reason := req.GetResolveReason()
+	state := req.GetState()
+	var tag pd.ResolutionTag
+
+	switch state {
+	case types.RequestState_APPROVED:
+		tag = pd.ResolvedApproved
+	case types.RequestState_DENIED:
+		tag = pd.ResolvedDenied
+	default:
+		logger.Get(ctx).Warningf("Unknown state %v (%s)", state, state.String())
+		return replyErr
+	}
+	err := trace.Wrap(a.updateMessages(ctx, req.GetName(), tag, reason, req.GetReviews()))
+	return trace.NewAggregate(replyErr, err)
+}
+
+func (a *BaseApp[T]) onDeletedRequest(ctx context.Context, reqID string) error {
+	return a.updateMessages(ctx, reqID, pd.ResolvedExpired, "", nil)
+}
+
+func (a *BaseApp[T]) broadcastMessages(ctx context.Context, channels []string, reqID string, reqData pd.AccessRequestData) error {
+	sentMessages, err := a.Bot.Broadcast(ctx, channels, reqID, reqData)
+	if len(sentMessages) == 0 && err != nil {
+		return trace.Wrap(err)
+	}
+	for _, data := range sentMessages {
+		logger.Get(ctx).WithFields(logger.Fields{
+			"slack_channel":   data.ChannelID,
+			"slack_timestamp": data.MessageID,
+		}).Info("Successfully posted to Slack")
+	}
+	if err != nil {
+		logger.Get(ctx).WithError(err).Error("Failed to post one or more messages to Slack")
+	}
+
+	_, err = a.pd.Update(ctx, reqID, func(existing GenericPluginData) (GenericPluginData, error) {
+		existing.SentMessages = sentMessages
+		return existing, nil
+	})
+
+	return trace.Wrap(err)
+}
+
+func (a *BaseApp[T]) postReviewReplies(ctx context.Context, reqID string, reqReviews []types.AccessReview) error {
+	var oldCount int
+
+	pd, err := a.pd.Update(ctx, reqID, func(existing GenericPluginData) (GenericPluginData, error) {
+		sentMessages := existing.SentMessages
+		if len(sentMessages) == 0 {
+			// wait for the plugin data to be updated with SentMessages
+			return GenericPluginData{}, trace.CompareFailed("existing sentMessages is empty")
+		}
+
+		count := len(reqReviews)
+		oldCount = existing.ReviewsCount
+		if oldCount >= count {
+			return GenericPluginData{}, trace.AlreadyExists("reviews are sent already")
+		}
+
+		existing.ReviewsCount = count
+		return existing, nil
+	})
+	if trace.IsAlreadyExists(err) {
+		logger.Get(ctx).Debug("Failed to post reply: replies are already sent")
+		return nil
+	}
+	if err != nil {
+		return trace.Wrap(err)
+	}
+
+	slice := reqReviews[oldCount:]
+	if len(slice) == 0 {
+		return nil
+	}
+
+	errors := make([]error, 0, len(slice))
+	for _, data := range pd.SentMessages {
+		ctx, _ = logger.WithFields(ctx, logger.Fields{"slack_channel": data.ChannelID, "slack_timestamp": data.MessageID})
+		for _, review := range slice {
+			if err := a.Bot.PostReviewReply(ctx, data.ChannelID, data.MessageID, review); err != nil {
+				errors = append(errors, err)
+			}
+		}
+	}
+	return trace.NewAggregate(errors...)
+}
+
+func (a *BaseApp[T]) getMessageRecipients(ctx context.Context, req types.AccessRequest) []string {
+	log := logger.Get(ctx)
+
+	// We receive a set from GetRecipientsFor but we still might end up with duplicate channel names.
+	// This can happen if this set contains the channel `C` and the email for channel `C`.
+	channelSet := stringset.New()
+
+	validEmaislSuggReviewers := []string{}
+	for _, reviewer := range req.GetSuggestedReviewers() {
+		if !lib.IsEmail(reviewer) {
+			log.Warningf("Failed to notify a suggested reviewer: %q does not look like a valid email", reviewer)
+			continue
+		}
+
+		validEmaislSuggReviewers = append(validEmaislSuggReviewers, reviewer)
+	}
+
+	recipients := a.Conf.GetRecipients().GetRecipientsFor(req.GetRoles(), validEmaislSuggReviewers)
+	for _, recipient := range recipients {
+		channel, err := a.Bot.FetchRecipient(ctx, recipient)
+		if err != nil {
+			// Something wrong happened, we log the error and continue to treat valid recipients
+			log.Warning(err)
+		} else {
+			channelSet.Add(channel.ID)
+		}
+	}
+
+	return channelSet.ToSlice()
+}
+
+// updateMessages updates the messages status and adds the resolve reason.
+func (a *BaseApp[T]) updateMessages(ctx context.Context, reqID string, tag pd.ResolutionTag, reason string, reviews []types.AccessReview) error {
+	log := logger.Get(ctx)
+
+	pluginData, err := a.pd.Update(ctx, reqID, func(existing GenericPluginData) (GenericPluginData, error) {
+		if len(existing.SentMessages) == 0 {
+			return GenericPluginData{}, trace.NotFound("plugin data not found")
+		}
+
+		// If resolution field is not empty then we already resolved the incident before. In this case we just quit.
+		if existing.AccessRequestData.ResolutionTag != pd.Unresolved {
+			return GenericPluginData{}, trace.CompareFailed("request is already resolved")
+		}
+
+		// Mark plugin data as resolved.
+		existing.ResolutionTag = tag
+		existing.ResolutionReason = reason
+
+		return existing, nil
+	})
+	if trace.IsNotFound(err) {
+		log.Debug("Failed to update messages: plugin data is missing")
+		return nil
+	}
+	if err != nil {
+		return trace.Wrap(err)
+	}
+
+	reqData, sentMessages := pluginData.AccessRequestData, pluginData.SentMessages
+	if err := a.Bot.UpdateMessages(ctx, reqID, reqData, sentMessages, reviews); err != nil {
+		return trace.Wrap(err)
+	}
+
+	log.Infof("Successfully marked request as %s in all messages", tag)
+
+	return nil
+}
