+++ access/discord/config.go
@@ -1,36 +1,34 @@
 package main
 
 import (
-	"strings"
-
 	"github.com/gravitational/teleport-plugins/access/common"
 	"github.com/gravitational/teleport-plugins/lib"
 	"github.com/gravitational/teleport/api/types"
 	"github.com/gravitational/trace"
 	"github.com/pelletier/go-toml"
+	"strings"
 )
 
-// SlackConfig stores the full configuration for the teleport-slack plugin to run.
-type SlackConfig struct {
+type DiscordConfig struct {
 	common.BaseConfig
-	Slack common.GenericAPIConfig
+	Discord common.GenericAPIConfig
 }
 
-// LoadSlackConfig reads the config file, initializes a new SlackConfig struct object, and returns it.
+// LoadDiscordConfig reads the config file, initializes a new DiscordConfig struct object, and returns it.
 // Optionally returns an error if the file is not readable, or if file format is invalid.
-func LoadSlackConfig(filepath string) (*SlackConfig, error) {
+func LoadDiscordConfig(filepath string) (*DiscordConfig, error) {
 	t, err := toml.LoadFile(filepath)
 	if err != nil {
 		return nil, trace.Wrap(err)
 	}
 
-	conf := &SlackConfig{}
+	conf := &DiscordConfig{}
 	if err := t.Unmarshal(conf); err != nil {
 		return nil, trace.Wrap(err)
 	}
 
-	if strings.HasPrefix(conf.Slack.Token, "/") {
-		conf.Slack.Token, err = lib.ReadPassword(conf.Slack.Token)
+	if strings.HasPrefix(conf.Discord.Token, "/") {
+		conf.Discord.Token, err = lib.ReadPassword(conf.Discord.Token)
 		if err != nil {
 			return nil, trace.Wrap(err)
 		}
@@ -45,11 +43,11 @@ func LoadSlackConfig(filepath string) (*SlackConfig, error) {
 // CheckAndSetDefaults checks the config struct for any logical errors, and sets default values
 // if some values are missing.
 // If critical values are missing and we can't set defaults for them — this will return an error.
-func (c *SlackConfig) CheckAndSetDefaults() error {
+func (c *DiscordConfig) CheckAndSetDefaults() error {
 	if err := c.Teleport.CheckAndSetDefaults(); err != nil {
 		return trace.Wrap(err)
 	}
-	if c.Slack.Token == "" {
+	if c.Discord.Token == "" {
 		return trace.BadParameter("missing required value slack.token")
 	}
 	if c.Log.Output == "" {
@@ -59,13 +57,13 @@ func (c *SlackConfig) CheckAndSetDefaults() error {
 		c.Log.Severity = "info"
 	}
 
-	if len(c.Slack.Recipients) > 0 {
+	if len(c.Discord.Recipients) > 0 {
 		if len(c.Recipients) > 0 {
 			return trace.BadParameter("provide either slack.recipients or role_to_recipients, not both.")
 		}
 
 		c.Recipients = common.RecipientsMap{
-			types.Wildcard: c.Slack.Recipients,
+			types.Wildcard: c.Discord.Recipients,
 		}
 	}
