+++ access/discord/version.go
@@ -1,0 +1,25 @@
+/*
+Copyright 2015-2021 Gravitational, Inc.
+
+Licensed under the Apache License, Version 2.0 (the "License");
+you may not use this file except in compliance with the License.
+You may obtain a copy of the License at
+
+    http://www.apache.org/licenses/LICENSE-2.0
+
+Unless required by applicable law or agreed to in writing, software
+distributed under the License is distributed on an "AS IS" BASIS,
+WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+See the License for the specific language governing permissions and
+limitations under the License.
+*/
+
+package main
+
+var (
+	// Version package version, specified in Makefile using ldflags
+	Version = `Not specified, use --ldflags "-X main.Version "1.0.0""`
+
+	// Gitref variable is specified in Makefile using ldflags
+	Gitref string
+)
