+++ access/email/config_test.go
@@ -21,7 +21,7 @@ import (
 	"path/filepath"
 	"testing"
 
-	"github.com/gravitational/teleport-plugins/access/config"
+	"github.com/gravitational/teleport-plugins/access/common"
 	"github.com/gravitational/teleport/api/types"
 	"github.com/gravitational/trace"
 	"github.com/stretchr/testify/require"
@@ -33,7 +33,7 @@ func TestRecipients(t *testing.T) {
 		desc             string
 		in               string
 		expectErr        require.ErrorAssertionFunc
-		expectRecipients config.RecipientsMap
+		expectRecipients common.RecipientsMap
 	}{
 		{
 			desc: "test delivery recipients",
@@ -45,7 +45,7 @@ func TestRecipients(t *testing.T) {
             sender = "email@example.org"
 			recipients = ["email1@example.org","email2@example.org"]
 			`,
-			expectRecipients: config.RecipientsMap{
+			expectRecipients: common.RecipientsMap{
 				types.Wildcard: []string{"email1@example.org", "email2@example.org"},
 			},
 		},
@@ -62,7 +62,7 @@ func TestRecipients(t *testing.T) {
 			"dev" = ["dev@example.org","sre@example.org"]
 			"*" = "admin@example.org"
 			`,
-			expectRecipients: config.RecipientsMap{
+			expectRecipients: common.RecipientsMap{
 				"dev":          []string{"dev@example.org", "sre@example.org"},
 				types.Wildcard: []string{"admin@example.org"},
 			},
